removed unitaustin namespace
changed GetDamage input value to float

+++ Freelancing Project/Assets/Game/Scripts/Interfaces/Idamageable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface Idamageable
 {
-    void GetDamaged(int damage);
+    void GetDamaged(float damage);
 }
